refactor auth token and separated out logic to be more utilizable
