+++ src/utils/jwt/index.ts
@@ -1,1 +1,1 @@
-export * from "./jwt"
+export * from "./jwt-token-manager";
